// utils/enums.ts
export enum Roles {
  ADMIN = "admin",
  MODERATOR = "moderator",
  STUDENT = "student",
}

export enum CourseStatus {
  DRAFT = "draft",
  PUBLISHED = "published",
  ARCHIVED = "archived",
}

export enum CourseLevel {
  BEGINNER = "beginner",
  INTERMEDIATE = "intermediate",
  ADVANCED = "advanced",
}

export enum EnrollmentStatus {
  ENROLLED = "enrolled",
  COMPLETED = "completed",
  WITHDRAWN = "withdrawn",
}

export enum PaymentStatus {
  PENDING = "pending",
  PAID = "paid",
  FAILED = "failed",
}